Refactor targeting helpers into separate module
